Simplify taxonomy json loading and put utr.json in a more logical place
